Adds migration of RPM Distributions to DistributionTrees
This patch comes with a new migration that adds pulp3_repository field to the Pulp2Repository
model. This is needed for creating a Variant associated with a DistributionTree in Pulp 3.

The Addons are completely ignored because pulp 2 did not support them.

Variants are only limited to the ones that reference the top level repository ('.').

closes: #6201
https://pulp.plan.io/issues/6201