Fix a bug with multiple copies of an errata end up in a repo
The packagelist of an errata depends on content within the repo. When an
errata is copied from one Pulp 2 repository to another, it is the same
content unit but with a different package list, but in Pulp 3 these
would be considered separate content units.

Because of this, when content is migrated, two Pulp 3 UpdateRecords are
created from the same Pulp2Content. Later when content units are
associated with the repositories, no accounting was being taken for the
fact that this could be the case, and that we need to pick the
particular content unit of the form that existed in the Pulp 2
repository rather than all of them.

closes: #7165
https://pulp.plan.io/issues/7165
(cherry picked from commit a29b2f08890c41be0de90709cd772e778d96e019)