Fix a bug where errata were not always migrated for new repositories
If an errata has already been migrated and is then added to a new
repository, which is then migrated, the new repository does not contain
the affected errata.

Ensure that new Pulp2Content are created when existing content are
migrated into new repos for the errata corner case (needs one per each repo).

closes: #7092
https://pulp.plan.io/issues/7092