Add the pulp2_subid field to the Pulp2Content model
fixes #8137
https://pulp.plan.io/issues/8137

The intended use case is to provide one to many mapping from Pulp 2 to
Pulp 3 content. This is achieved by creating multiple Pulp2Content
entries that all share a single pulp2_id, and differ only via the
pulp2_subid.

This is needed for structured Debian migrations:
https://pulp.plan.io/issues/7865

The extra Pulp2Content subrecords are added by the relevant
pre_migrate_content_detail method. The reason is that it is only known
what subrecords are needed once the content is actually inspected by
this method.