Remove a test which attemps to check that only one reset can be run
The problem is that with the small amount of data which is used
for tests in CI, reset happens very quickly and very often is finished
before we are able to trigger the second task to test its failure.

The benefit of this test is low, the disturbance from regular test failures
because the first task is completed very quickly is high.

[noissue]