Use proper 2to3 uniqueness constraints for deb
closes #484
https://github.com/pulp/pulp-2to3-migration/issues/484

Previously the uniqueness constraints allowed for only one 2to3 unit per
Pulp 3 unit to be created. This causes problems when there is more than
one Pulp 2 unit that map onto the same Pulp 3 unit to be created.
If prevents the resultant Pulp 3 unit from being added to every repo it
belongs in.